+++ modules/apps/transfer/keeper/relay_test.go
@@ -62,6 +62,15 @@ func (suite *KeeperTestSuite) TestSendTransfer() {
 			}, true,
 		},
 		{
+			"successful transfer of entire balance",
+			func() {
+				coin.Amount = types.UnboundedSpendLimit()
+				var ok bool
+				expEscrowAmount, ok = sdkmath.NewIntFromString(ibctesting.DefaultGenesisAccBalance)
+				suite.Require().True(ok)
+			}, true,
+		},
+		{
 			"source channel not found",
 			func() {
 				// channel references wrong ID