+++ modules/apps/transfer/types/coin.go
@@ -2,12 +2,16 @@ package types
 
 import (
 	"fmt"
+	"math/big"
 	"strings"
 
 	sdkmath "cosmossdk.io/math"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
+
+// maxUint256 is the maximum value for a 256 bit unsigned integer.
+var maxUint256 = new(big.Int).Sub(new(big.Int).Lsh(big.NewInt(1), 256), big.NewInt(1))
 
 // SenderChainIsSource returns false if the denomination originally came
 // from the receiving chain and true otherwise.
@@ -46,4 +50,13 @@ func GetPrefixedDenom(portID, channelID, baseDenom string) string {
 func GetTransferCoin(portID, channelID, baseDenom string, amount sdkmath.Int) sdk.Coin {
 	denomTrace := ParseDenomTrace(GetPrefixedDenom(portID, channelID, baseDenom))
 	return sdk.NewCoin(denomTrace.IBCDenom(), amount)
+}
+
+// UnboundedSpendLimit returns the sentinel value that can be used
+// as the amount for a denomination's spend limit for which spend limit updating
+// should be disabled. Please note that using this sentinel value means that a grantee
+// will be granted the privilege to do ICS20 token transfers for the total amount
+// of the denomination available at the granter's account.
+func UnboundedSpendLimit() sdkmath.Int {
+	return sdkmath.NewIntFromBigInt(maxUint256)
 }